requirements.txt
fastapi
uvicorn[standard]
langchain
langchain-community
faiss-cpu
openai
sqlalchemy
pydantic
python-dotenv
